+++ crypt-core/src/test/java/de/alpharogroup/crypto/file/FileEncryptDecryptorTest.java
@@ -24,22 +24,53 @@
  */
 package de.alpharogroup.crypto.file;
 
+import static org.testng.AssertJUnit.assertEquals;
+
 import java.io.File;
+import java.io.FileNotFoundException;
 
 import javax.crypto.Cipher;
 
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import de.alpharogroup.AbstractTestCase;
+import de.alpharogroup.crypto.algorithm.MdAlgorithm;
 import de.alpharogroup.crypto.algorithm.SunJCEAlgorithm;
 import de.alpharogroup.crypto.model.CryptModel;
+import de.alpharogroup.file.checksum.ChecksumExtensions;
 import de.alpharogroup.file.delete.DeleteFileExtensions;
 import de.alpharogroup.file.search.PathFinder;
 
 /**
  * The unit test class for the class {@link FileEncryptor} and the class {@link FileDecryptor}
  */
-public class FileEncryptDecryptorTest
+public class FileEncryptDecryptorTest extends AbstractTestCase<String, String>
 {
+
+	File cryptDir;
+	CryptModel<Cipher, String> cryptModel;
+	File decrypted;
+	FileDecryptor decryptor;
+	File dirToEncrypt;
+	File encrypted;
+	FileEncryptor encryptor;
+	String firstKey;
+	File toEncrypt;
+
+	/**
+	 * Sets up method will be invoked before every unit test method in this class
+	 */
+	@BeforeMethod
+	protected void setUp()
+	{
+		cryptDir = new File(PathFinder.getSrcTestResourcesDir(), "crypt");
+		toEncrypt = new File(cryptDir, "test.txt");
+		dirToEncrypt = new File(cryptDir, "food");
+		firstKey = "D1D15ED36B887AF1";
+		cryptModel = CryptModel.<Cipher, String> builder().key(firstKey)
+			.algorithm(SunJCEAlgorithm.PBEWithMD5AndDES).build();
+	}
 
 	/**
 	 * Test method for the encrpytion with the class {@link FileEncryptor} and decryption with the
@@ -51,25 +82,36 @@ public class FileEncryptDecryptorTest
 	@Test
 	public void testEncryptDecryptConstructorFiles() throws Exception
 	{
-		final File cryptDir = new File(PathFinder.getSrcTestResourcesDir(), "crypt");
-		final File toEncrypt = new File(cryptDir, "test.txt");
+		// new scenario...
+		encryptor = new FileEncryptor(cryptModel, new File(cryptDir, "encryptedCnstr.enc"));
+		encrypted = encryptor.encrypt(toEncrypt);
 
-		final String firstKey = "D1D15ED36B887AF1";
+		decryptor = new FileDecryptor(cryptModel, new File(cryptDir, "decryptedCnstr.decrypted"));
 
-		final CryptModel<Cipher, String> cryptModel = CryptModel.<Cipher, String> builder()
-			.key(firstKey).algorithm(SunJCEAlgorithm.PBEWithMD5AndDES).build();
+		decrypted = decryptor.decrypt(encrypted);
 
-		final FileEncryptor encryptor = new FileEncryptor(cryptModel,
-			new File(cryptDir, "encryptedCnstr.enc"));
-		final File encrypted = encryptor.encrypt(toEncrypt);
-
-		final FileDecryptor decryptor = new FileDecryptor(cryptModel,
-			new File(cryptDir, "decryptedCnstr.decrypted"));
-
-		final File decrypted = decryptor.decrypt(encrypted);
+		expected = ChecksumExtensions.getChecksum(toEncrypt, MdAlgorithm.MD5);
+		actual = ChecksumExtensions.getChecksum(decrypted, MdAlgorithm.MD5);
+		assertEquals(actual, expected);
 		// clean up...
 		DeleteFileExtensions.delete(encrypted);
 		DeleteFileExtensions.delete(decrypted);
+	}
+
+	/**
+	 * Test method for the encrpytion with the class {@link FileEncryptor} and decryption with the
+	 * class {@link FileDecryptor} with given constructor files that throws a
+	 * {@link FileNotFoundException}
+	 *
+	 * @throws Exception
+	 *             is thrown if any error occurs on the execution
+	 */
+	@Test(expectedExceptions = FileNotFoundException.class)
+	public void testEncryptDecryptConstructorFilesThrowFileNotFoundException() throws Exception
+	{
+		// new scenario...
+		encryptor = new FileEncryptor(cryptModel, new File(cryptDir, "foodenc"));
+		encrypted = encryptor.encrypt(dirToEncrypt);
 	}
 
 	/**
@@ -82,20 +124,16 @@ public class FileEncryptDecryptorTest
 	@Test
 	public void testEncryptDecryptDefaultFiles() throws Exception
 	{
-		final File cryptDir = new File(PathFinder.getSrcTestResourcesDir(), "crypt");
-		final File toEncrypt = new File(cryptDir, "test.txt");
+		encryptor = new FileEncryptor(cryptModel);
+		encrypted = encryptor.encrypt(toEncrypt);
 
-		final String firstKey = "D1D15ED36B887AF1";
+		decryptor = new FileDecryptor(cryptModel);
 
-		final CryptModel<Cipher, String> cryptModel = CryptModel.<Cipher, String> builder()
-			.key(firstKey).algorithm(SunJCEAlgorithm.PBEWithMD5AndDES).build();
+		decrypted = decryptor.decrypt(encrypted);
 
-		final FileEncryptor encryptor = new FileEncryptor(cryptModel);
-		final File encrypted = encryptor.encrypt(toEncrypt);
-
-		final FileDecryptor decryptor = new FileDecryptor(cryptModel);
-
-		final File decrypted = decryptor.decrypt(encrypted);
+		expected = ChecksumExtensions.getChecksum(toEncrypt, MdAlgorithm.MD5);
+		actual = ChecksumExtensions.getChecksum(decrypted, MdAlgorithm.MD5);
+		assertEquals(actual, expected);
 		// clean up...
 		DeleteFileExtensions.delete(encrypted);
 		DeleteFileExtensions.delete(decrypted);
@@ -111,15 +149,7 @@ public class FileEncryptDecryptorTest
 	@Test
 	public void testEncryptDecryptFactoryInjected() throws Exception
 	{
-		final File cryptDir = new File(PathFinder.getSrcTestResourcesDir(), "crypt");
-		final File toEncrypt = new File(cryptDir, "test.txt");
-
-		final String firstKey = "D1D15ED36B887AF1";
-
-		final CryptModel<Cipher, String> cryptModel = CryptModel.<Cipher, String> builder()
-			.key(firstKey).algorithm(SunJCEAlgorithm.PBEWithMD5AndDES).build();
-
-		final FileEncryptor encryptor = new FileEncryptor(cryptModel)
+		encryptor = new FileEncryptor(cryptModel)
 		{
 
 			/** The Constant serialVersionUID. */
@@ -131,9 +161,9 @@ public class FileEncryptDecryptorTest
 				return new File(cryptDir, "encryptedFctrNjctd.enc");
 			}
 		};
-		final File encrypted = encryptor.encrypt(toEncrypt);
+		encrypted = encryptor.encrypt(toEncrypt);
 
-		final FileDecryptor decryptor = new FileDecryptor(cryptModel)
+		decryptor = new FileDecryptor(cryptModel)
 		{
 
 			/** The Constant serialVersionUID. */
@@ -146,7 +176,12 @@ public class FileEncryptDecryptorTest
 			}
 		};
 
-		final File decrypted = decryptor.decrypt(encrypted);
+		decrypted = decryptor.decrypt(encrypted);
+
+		expected = ChecksumExtensions.getChecksum(toEncrypt, MdAlgorithm.MD5);
+		actual = ChecksumExtensions.getChecksum(decrypted, MdAlgorithm.MD5);
+		assertEquals(actual, expected);
+
 		// clean up...
 		DeleteFileExtensions.delete(encrypted);
 		DeleteFileExtensions.delete(decrypted);