+++ crypt-data/src/main/java/de/alpharogroup/crypto/CryptConst.java
@@ -30,6 +30,7 @@ import de.alpharogroup.crypto.algorithm.KeyPairGeneratorAlgorithm;
 import de.alpharogroup.crypto.algorithm.MacAlgorithm;
 import de.alpharogroup.crypto.algorithm.MdAlgorithm;
 import de.alpharogroup.crypto.algorithm.SunJCEAlgorithm;
+import de.alpharogroup.crypto.algorithm.UnionWord;
 import de.alpharogroup.crypto.mechanisms.PBEMechanism;
 import de.alpharogroup.crypto.pw.PasswordHashType;
 
@@ -42,9 +43,6 @@ import de.alpharogroup.crypto.pw.PasswordHashType;
 public abstract class CryptConst
 {
 
-	/** The Constant AND. */
-	public static final String AND = "And";
-
 	/**
 	 * Constant for the utf-8 encoding.
 	 */
@@ -55,26 +53,20 @@ public abstract class CryptConst
 	 */
 	public static final int ITERATIONCOUNT = 19;
 
-	/** The Constant PBE. */
-	public static final String PBE = PBEMechanism.PBE.name();
-
-	/** The Constant WITH. */
-	public static final String WITH = "With";
-
 	/** The Constant PBE_WITH. */
-	public static final String PBE_WITH = PBE + WITH;
+	public static final String PBE_WITH = PBEMechanism.PBE.name() + UnionWord.With.name();
 
 	/**
 	 * Constant for the algorithm 'PBEWithMD5AndAES' to encrypt and decrypt.
 	 */
-	public static final String PBE_WITH_MD5_AND_AES = PBE_WITH + MdAlgorithm.MD5.name() + AND
-		+ AesAlgorithm.AES.name();
+	public static final String PBE_WITH_MD5_AND_AES = PBE_WITH + MdAlgorithm.MD5.name()
+		+ UnionWord.And.name() + AesAlgorithm.AES.name();
 
 	/**
 	 * Constant for the algorithm 'PBEWithMD5AndDES' to encrypt and decrypt.
 	 */
-	public static final String PBE_WITH_MD5_AND_DES = PBE_WITH + MdAlgorithm.MD5.name() + AND
-		+ SunJCEAlgorithm.DES.name();
+	public static final String PBE_WITH_MD5_AND_DES = PBE_WITH + MdAlgorithm.MD5.name()
+		+ UnionWord.And.name() + SunJCEAlgorithm.DES.name();
 
 	/**
 	 * Constant for the algorithm 'PBEWITHSHA1AND128BITAES-CBC-BC' to encrypt and decrypt.
@@ -85,7 +77,7 @@ public abstract class CryptConst
 	 * Constant for the algorithm 'PBEWithSHA1AndDESede' to encrypt and decrypt.
 	 */
 	public static final String PBE_WITH_SHA1_AND_DES_EDE = PBE_WITH + HashAlgorithm.SHA1.name()
-		+ AND + SunJCEAlgorithm.DESede.name();
+		+ UnionWord.And.name() + SunJCEAlgorithm.DESede.name();
 
 	/** The Constant PBKDF2. */
 	public static final String PBKDF2 = PasswordHashType.PBKDF2.name();
@@ -93,8 +85,8 @@ public abstract class CryptConst
 	/**
 	 * Constant for the algorithm 'PBKDF2WithHmacSHA1' to encrypt and decrypt.
 	 */
-	public static final String PBKDF2_WITH_HMAC_SHA1 = PBKDF2 + WITH + MacAlgorithm.HMAC
-		+ HashAlgorithm.SHA1.name();
+	public static final String PBKDF2_WITH_HMAC_SHA1 = PBKDF2 + UnionWord.With.name()
+		+ MacAlgorithm.HMAC + HashAlgorithm.SHA1.name();
 
 	/**
 	 * Constant for the private key.
@@ -109,7 +101,7 @@ public abstract class CryptConst
 
 	/** Constant for algorithm 'SHA256withRSA' to encrypt and decrypt. */
 	public static final String SHA256_WITH_RSA = HashAlgorithm.SHA256.getAlgorithm()
-		+ CryptConst.WITH.toLowerCase() + KeyPairGeneratorAlgorithm.RSA.getAlgorithm();
+		+ UnionWord.With.name().toLowerCase() + KeyPairGeneratorAlgorithm.RSA.getAlgorithm();
 
 	/**
 	 * Constant for the initialization from SimpleDecryptor.