+++ crypt-data/src/test/java/de/alpharogroup/crypto/factories/CertFactoryTest.java
@@ -54,9 +54,9 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.crypto.CryptConst;
 import de.alpharogroup.crypto.algorithm.HashAlgorithm;
 import de.alpharogroup.crypto.algorithm.KeyPairGeneratorAlgorithm;
+import de.alpharogroup.crypto.algorithm.UnionWord;
 import de.alpharogroup.crypto.key.reader.CertificateReader;
 import de.alpharogroup.crypto.key.reader.PrivateKeyReader;
 import de.alpharogroup.crypto.key.reader.PublicKeyReader;
@@ -197,8 +197,8 @@ public class CertFactoryTest
 
 		final String subject = "CN=Test subject";
 		final String issuer = "CN=Test issue";
-		final String signatureAlgorithm = HashAlgorithm.SHA256.getAlgorithm() + CryptConst.WITH
-			+ KeyPairGeneratorAlgorithm.RSA.getAlgorithm();
+		final String signatureAlgorithm = HashAlgorithm.SHA256.getAlgorithm()
+			+ UnionWord.With.name() + KeyPairGeneratorAlgorithm.RSA.getAlgorithm();
 
 		final Date start = Date.from(
 			LocalDate.of(2017, Month.JANUARY, 1).atStartOfDay(ZoneId.systemDefault()).toInstant());